- code-gen: replaces main client with generated proxy
