+++ packages/test/generatedCode/apiClient.ts
@@ -2,7 +2,7 @@
 // eslint-disable
 // Generated by Microsoft Kiota
 import { type UsersRequestBuilder, UsersRequestBuilderNavigationMetadata, UsersRequestBuilderUriTemplate } from './users/';
-import { enableBackingStoreForSerializationWriterFactory, ParseNodeFactoryRegistry, registerDefaultDeserializer, registerDefaultSerializer, SerializationWriterFactoryRegistry, type BaseRequestBuilder, type KeysToExcludeForNavigationMetadata, type NavigationMetadata, type RequestAdapter, type RequestMetadata } from '@microsoft/kiota-abstractions';
+import { apiClientProxifier, enableBackingStoreForSerializationWriterFactory, ParseNodeFactoryRegistry, registerDefaultDeserializer, registerDefaultSerializer, SerializationWriterFactoryRegistry, type BaseRequestBuilder, type KeysToExcludeForNavigationMetadata, type NavigationMetadata, type RequestAdapter, type RequestMetadata } from '@microsoft/kiota-abstractions';
 import { FormParseNodeFactory, FormSerializationWriterFactory } from '@microsoft/kiota-serialization-form';
 import { JsonParseNodeFactory, JsonSerializationWriterFactory } from '@microsoft/kiota-serialization-json';
 import { MultipartSerializationWriterFactory } from '@microsoft/kiota-serialization-multipart';
@@ -16,6 +16,22 @@ export interface ApiClient extends BaseRequestBuilder<ApiClient> {
      * The users property
      */
     get users(): UsersRequestBuilder;
+}
+export function newApiClient(requestAdapter: RequestAdapter) {
+    registerDefaultSerializer(JsonSerializationWriterFactory);
+    registerDefaultSerializer(TextSerializationWriterFactory);
+    registerDefaultSerializer(FormSerializationWriterFactory);
+    registerDefaultSerializer(MultipartSerializationWriterFactory);
+    registerDefaultDeserializer(JsonParseNodeFactory);
+    registerDefaultDeserializer(TextParseNodeFactory);
+    registerDefaultDeserializer(FormParseNodeFactory);
+    if (requestAdapter.baseUrl === undefined || requestAdapter.baseUrl === "") {
+        requestAdapter.baseUrl = "https://graph.microsoft.com/v1.0";
+    }
+    const pathParameters: Record<string, unknown> = {
+        "baseurl": requestAdapter.baseUrl,
+    };
+    return apiClientProxifier<ApiClient>(requestAdapter, pathParameters, ApiClientUriTemplate, ApiClientNavigationMetadata, undefined);
 }
 export const ApiClientNavigationMetadata: Record<Exclude<keyof ApiClient, KeysToExcludeForNavigationMetadata>, NavigationMetadata> = {
     "users": {
