+++ packages/test/generatedCode/kiota-lock.json
@@ -1,5 +1,5 @@
 {
-  "descriptionHash": "293EC13E4EA89E6695FB19ADCE23A67734E8CC8018E1BA88D27E27EC369C01F1C6EA5C84B7F25E60128993F20D175018DDDB4BA206CB4A7336F09CB8352B74C3",
+  "descriptionHash": "23E4EA89E6695FB19ADCE23A67734E8CC8018E1BA88D27E27EC369C01F1C6EA5C84B7F25E60128993F20D175018DDDB4BA206CB4A7336F09CB8352B74C3",
   "descriptionLocation": "https://raw.githubusercontent.com/microsoftgraph/msgraph-sdk-powershell/dev/openApiDocs/v1.0/Mail.yml",
   "lockFileVersion": "1.0.0",
   "kiotaVersion": "1.10.0",
