+++ packages/test/tests/testClient.ts
@@ -1,12 +1,12 @@
 import {FetchRequestAdapter} from "@microsoft/kiota-http-fetchlibrary";
 import {AzureIdentityAuthenticationProvider} from "@microsoft/kiota-authentication-azure";
 import {ClientSecretCredential} from "@azure/identity";
-import { getNewApiClient } from "../generatedCode/apiClientProxy";
+import { newApiClient } from "../generatedCode/apiClient";
 
 const tokenCredential = new ClientSecretCredential(process.env.TENANT_ID!, process.env.CLIENT_ID!, process.env.CLIENT_SECRET! );
 
 const authProvider = new AzureIdentityAuthenticationProvider(tokenCredential, ["https://graph.microsoft.com/.default"]);
 const fetchRequestAdapter = new FetchRequestAdapter(authProvider)
-export const proxyClient = getNewApiClient(fetchRequestAdapter);
+export const proxyClient = newApiClient(fetchRequestAdapter);
 export const userId = process.env.USER_ID!;
